move logic up to adv-post-form
