+++ components/job-form.js
@@ -19,6 +19,7 @@ import { MAX_TITLE_LENGTH, MEDIA_URL } from '@/lib/constants'
 import { useToast } from './toast'
 import { toastDeleteScheduled } from '@/lib/form'
 import { ItemButtonBar } from './post'
+import { useFormikContext } from 'formik'
 
 function satsMin2Mo (minute) {
   return minute * 30 * 24 * 60
@@ -166,6 +167,8 @@ export default function JobForm ({ item, sub }) {
 }
 
 function PromoteJob ({ item, sub }) {
+  const formik = useFormikContext()
+  const [show, setShow] = useState(undefined)
   const [monthly, setMonthly] = useState(satsMin2Mo(item?.maxBid || 0))
   const [getAuctionPosition, { data }] = useLazyQuery(gql`
     query AuctionPosition($id: ID, $bid: Int!) {
@@ -180,10 +183,26 @@ function PromoteJob ({ item, sub }) {
     setMonthly(satsMin2Mo(initialMaxBid))
   }, [])
 
+  useEffect(() => {
+    const isDirty = !!formik?.values.maxBid && (formik?.maxBid !== 0 || formik?.maxBid !== '')
+
+    if (isDirty) {
+      setShow(false)
+    }
+  }, [formik?.values])
+
+  useEffect(() => {
+    const hasMaxBidError = !!formik?.errors?.maxBid && formik?.errors?.maxBid !== ''
+
+    if (formik?.isSubmitting) {
+      setShow(hasMaxBidError ? !show : undefined)
+    }
+
+  }, [formik?.isSubmitting])
+
   return (
     <AccordianItem
-      hasForm
-      show={item?.maxBid > 0}
+      show={item?.maxBid > 0 || show}
       header={<div style={{ fontWeight: 'bold', fontSize: '92%' }}>promote</div>}
       body={
         <>
